Add perl libGD modifications
Perl's GD library fails to compiled and install automatically due to the
presence of unwritable files in the `bdf_scripts/` folder. This is an
issue known to upstream:
https://rt.cpan.org/Public/Bug/Display.html?id=99901 but left unfixed
despite rendering the package uninstallable from source.

Thus, this docker build script exists to make those scripts writeable
and thus render the package installable, so it can be used in the circos
installation definition. https://github.com/galaxyproject/tools-iuc/issues/77