Added check-for-updates script
This is an EXTREMELY experimental tool. Checking for updates is a waste
of time. We should keep up with updates to the software we care about.
Thus, `check-for-updates` was born.

It inspects the 'default' entry of every package with a `build.yml`
file. If that's found, then it extracts the urls from the urls section,
and checks for those with a `${version}` in them. When those are found,
it uses some hardcoded logic to check for more recent releases. When a
more recent release is found, it will copy the build.yml file across,
and generate a new folder/version with the updated version number value.

This should let the docker image for the original be run once, that
archive generated. Later, updates can be checked for, and any missing
packages can be generated.